requirements.txt
python-dotenv==1.0.1
numpy==2.0.1
faiss-cpu==1.8.0
google-generativeai==0.7.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
pydantic==2.8.2
sqlglot==24.4.0
langgraph==0.2.27
langchain-core==0.2.21
typing_extensions==4.12.2
pyyaml==6.0.2

